% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{ambiguous}
\alias{ambiguous}
\title{Test if a given numeric vector can have different meanings.}
\usage{
ambiguous(digits)
}
\arguments{
\item{digits}{A numeric vector.}
}
\value{
Logical. True if the number vector has multiple interpretations.
}
\description{
Test if a given numeric vector can have different meanings.
}
\keyword{internal}
