% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{digits_from}
\alias{digits_from}
\title{Generate digits from text with spelt numbers in a supported language.}
\usage{
digits_from(text, lang = "en")
}
\arguments{
\item{text}{word(s) that spell numbers}

\item{lang}{the language of the numbers currently en|fr|es}
}
\value{
numeric vector
}
\description{
Generate digits from text with spelt numbers in a supported language.
}
\examples{
\dontrun{
digits_from("five hundred and thirty eight")
}

}
