% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{.numberize}
\alias{.numberize}
\title{Internal function used in the numberize() call for vectors.}
\usage{
.numberize(text, lang = c("en", "fr", "es"))
}
\arguments{
\item{text}{Character string in a supported language.}

\item{lang}{Language of the character string.
Currently one of \code{"en" | "fr" | "es"}.}
}
\value{
A numeric value.
}
\description{
Internal function used in the numberize() call for vectors.
}
\keyword{internal}
