% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{number_from}
\alias{number_from}
\title{Generate number from a numeric vector.
This is the output of the digits_from()}
\usage{
number_from(digits)
}
\arguments{
\item{digits}{a numeric vector translated from words}
}
\value{
numeric vector
}
\description{
Generate number from a numeric vector.
This is the output of the digits_from()
}
\examples{
\dontrun{
number_from(c(500, 40, 5))
}

}
