% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{numberize}
\alias{numberize}
\title{Convert a text string of spelt numbers to its numeric equivalent.}
\usage{
numberize(text, lang = "en")
}
\arguments{
\item{text}{string containing spelt numbers in a supported language.}

\item{lang}{the language of the spelt numbers. Currently en|fr|es.}
}
\value{
numeric value

numeric vector
}
\description{
Convert a text string of spelt numbers to its numeric equivalent.
}
\examples{
\dontrun{
numberize("five hundred and thirty eight")
}

}
