% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{digits_from}
\alias{digits_from}
\title{Generate a numeric vector from text in a supported language.}
\usage{
digits_from(text, lang = "en")
}
\arguments{
\item{text}{Word(s) that spell numbers. e.g. "one", "deux", "trois"}

\item{lang}{The text's language. Currently one of \code{"en" | "fr" | "es"}}
}
\value{
A numeric vector.
}
\description{
Generate a numeric vector from text in a supported language.
}
\examples{
\dontrun{
digits_from("five hundred and thirty eight")
# [1]   5 100  30   8
}

}
