% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{number_from}
\alias{number_from}
\title{Generate number from a numeric vector.
Uses \code{digits_from()} output to generate the numeric value of the text.}
\usage{
number_from(digits)
}
\arguments{
\item{digits}{A numeric vector translated from words.}
}
\value{
A numeric value.
}
\description{
Generate number from a numeric vector.
Uses \code{digits_from()} output to generate the numeric value of the text.
}
\examples{
\dontrun{
number_from(c(5, 100, 30, 8))
# [1] 538
}

}
