% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{numberize}
\alias{numberize}
\title{Convert a vector string of spelled numbers in a supported language to
its numeric equivalent.}
\usage{
numberize(text, lang = c("en", "fr", "es"))
}
\arguments{
\item{text}{Vector containing spelled numbers in a supported language.}

\item{lang}{The text's language. Currently one of \code{"en" | "fr" | "es"}.
Default is "en"}
}
\value{
A vector of numeric values.
}
\description{
Convert a vector string of spelled numbers in a supported language to
its numeric equivalent.
}
\examples{
# convert to numbers a scalar
numberize("five hundred and thirty eight")

# convert a vector of values
numberize(c("dix", "soixante-cinq", "deux mille vingt-quatre"), lang = "fr")

}
